1.0 Separate Data from Code
First Major division between data and code
