+++ SU-MA-DO/SU-MA-DO.py
@@ -1,87 +1,7 @@
 from random import shuffle
 import timeit
 import itertools
-
-
-def get_adj_list(number):
-    triangle = {
-    	0:[3,1]
-    	,1: [0,2,3,4]
-    	,2: [1,4]
-    	,3: [0,1,4,5]
-    	,4: [1,2,3,5]
-    	,5: [3,4]
-    }
-    
-    rel2x3 = {
-        0: [1, 3, 4]
-        , 1: [0, 2, 4]
-        , 2: [1, 5]
-        , 3: [0, 4]
-        , 4: [0, 1, 3, 5]
-        , 5: [2, 4]
-    }
-
-    rel3x3 = {
-        0: [1, 3, 4]
-        , 1: [0, 2, 4]
-        , 2: [1, 5]
-        , 3: [0, 4, 6]
-        , 4: [0, 1, 3, 5, 7, 8]
-        , 5: [2, 4, 8]
-        , 6: [3, 7]
-        , 7: [4, 6, 8]
-        , 8: [4, 5, 7]
-    }
-
-    rel4x4 = {
-        0: [1, 4]
-        , 1: [0, 2, 4, 5]
-        , 2: [1, 3, 6, 7]
-        , 3: [2, 7]
-        , 4: [0, 1, 5, 8]
-        , 5: [1, 4, 6, 9]
-        , 6: [2, 5, 7, 10]
-        , 7: [2, 3, 6, 11]
-        , 8: [4, 9, 12, 13]
-        , 9: [5, 8, 10, 13]
-        , 10: [6, 9, 11, 14]
-        , 11: [7, 10, 14, 15]
-        , 12: [8, 13]
-        , 13: [8, 9, 12, 14]
-        , 14: [10, 11, 13, 15]
-        , 15: [11, 14]
-    }
-
-    rel5x5 = {
-        0: [1, 5]
-        , 1: [0, 2, 6]
-        , 2: [1, 3, 6, 8]
-        , 3: [2, 4, 8]
-        , 4: [3, 9]
-        , 5: [0, 6, 10]
-        , 6: [1, 2, 5, 7, 10, 11]
-        , 7: [6, 8, 12]
-        , 8: [2, 3, 7, 9, 13, 14]
-        , 9: [4, 8, 14]
-        , 10: [5, 6, 15, 16]
-        , 11: [6, 12, 16]
-        , 12: [7, 11, 13, 17]
-        , 13: [8, 12, 18]
-        , 14: [8, 9, 18, 19]
-        , 15: [10, 16, 20]
-        , 16: [10, 11, 15, 17, 21, 22]
-        , 17: [16, 12, 18]
-        , 18: [13, 14, 17, 19, 22, 23]
-        , 19: [14, 18, 24]
-        , 20: [15, 21]
-        , 21: [16, 20, 22]
-        , 22: [16, 18, 21, 23]
-        , 23: [18, 22, 24]
-        , 24: [19, 23]
-    }
-    listofrelations = {"t":triangle, "2": rel2x3, "3": rel3x3, "4": rel4x4, "5": rel5x5}
-    return listofrelations[number]
+from relations_dicts import get_adj_list
 
 
 def main(nro_rel=3, shuffledlist=None):
@@ -206,7 +126,6 @@ def generatePoligons(sl, rel):
     h = {sum(sl[j] for j in i) for i in rel.values()}
     return h
 
-
 """
 
 Testing
@@ -309,4 +228,5 @@ def testing():
         print("EXITO")
     else:
         print("FRACASO")
+
 testing()
